Next conig fixes
