+++ .eslintrc.json
@@ -3,7 +3,7 @@
     "browser": true,
     "es2021": true
   },
-  "extends": ["plugin:react/recommended", "standard-with-typescript"],
+  "extends": ["plugin:react/recommended", "standard-with-typescript", "next", "prettier"],
   "overrides": [],
   "parserOptions": {
     "ecmaVersion": "latest",
