+++ package.json
@@ -16,6 +16,7 @@
     "@nextui-org/react": "^2.1.13",
     "autoprefixer": "^10.4.16",
     "classnames": "^2.3.2",
+    "encoding": "^0.1.13",
     "express": "^4.17.1",
     "express-favicon": "^2.0.1",
     "framer-motion": "^10.16.4",
